OCR_ICD_Case_Study/requirements.txt

fastapi
uvicorn
pydantic
python-multipart
pillow
requests
langchain
langchain-openai
langchain-community
faiss-cpu
openai
psycopg2-binary
